Use TIMESTAMPTZ in schema, not TIMESTAMP
TIMESTAMP in Postgres is what LocalTime is in Haskell. TIMESTAMPTZ in
Postgres (or TIMESTAMP WITH TIME ZONE in long form) is what UTCTime is
in Haskell. Unlike what the name suggests, TIMESTAMPTZ is *not* what
ZonedTime is in Haskell.

While it is possible to use TIMESTAMP correctly, doing so is quite
subtle. If the intention is to record when something happened,
TIMESTAMPTZ is a safe default.

CURRENT_TIMESTAMP in Postgres is a TIMESTAMPTZ. If you use it to insert
into a TIMESTAMP column, Postgres casts it silently. In such as cast,
Postgres formats CURRENT_TIMESTAMP as a local time in the current time
zone.

The current time zone depends on the Postgres configuration and
the connection. I don't know what the default configuration ends up
being on NixOS, but on most distributions it will be set to the server's
local time zone. I don't know what the default local time zone is on
NixOS.

If the current time zone is set to a time zone that observes daylight
savings time, then the tweets created during the extra hour of a DST
change will not be ordered chronologically; the extra hour will be
interleaved with the preceding hour. Also, changing the current time
zone would require great care. Using TIMESTAMPTZ sidesteps both
problems.

See also:

 * https://www.postgresql.org/docs/12/datatype-datetime.html
 * https://www.postgresql.org/docs/12/runtime-config-client.html#GUC-TIMEZONE